refactor(backend): Handshake service
