+++ packages/backend/src/tokens/request.ts
@@ -15,7 +15,7 @@ import type { HandshakeState, RequestState, SignedInState, SignedOutState } from
 import { AuthErrorReason, handshake, signedIn, signedOut } from './authStatus';
 import { createClerkRequest } from './clerkRequest';
 import { getCookieName, getCookieValue } from './cookie';
-import { verifyHandshakeToken } from './handshake';
+import { HandshakeService } from './handshake';
 import type { AuthenticateRequestOptions, OrganizationSyncOptions } from './types';
 import { verifyToken } from './verify';
 
@@ -58,26 +58,6 @@ function assertSignInUrlFormatAndOrigin(_signInUrl: string, origin: string) {
   }
 }
 
-/**
- * Currently, a request is only eligible for a handshake if we can say it's *probably* a request for a document, not a fetch or some other exotic request.
- * This heuristic should give us a reliable enough signal for browsers that support `Sec-Fetch-Dest` and for those that don't.
- */
-function isRequestEligibleForHandshake(authenticateContext: { secFetchDest?: string; accept?: string }) {
-  const { accept, secFetchDest } = authenticateContext;
-
-  // NOTE: we could also check sec-fetch-mode === navigate here, but according to the spec, sec-fetch-dest: document should indicate that the request is the data of a user navigation.
-  // Also, we check for 'iframe' because it's the value set when a doc request is made by an iframe.
-  if (secFetchDest === 'document' || secFetchDest === 'iframe') {
-    return true;
-  }
-
-  if (!secFetchDest && accept?.startsWith('text/html')) {
-    return true;
-  }
-
-  return false;
-}
-
 function isRequestEligibleForRefresh(
   err: TokenVerificationError,
   authenticateContext: { refreshTokenInCookie?: string },
@@ -105,124 +85,8 @@ export async function authenticateRequest(
     assertProxyUrlOrDomain(authenticateContext.proxyUrl || authenticateContext.domain);
   }
 
-  // NOTE(izaak): compute regex matchers early for efficiency - they can be used multiple times.
   const organizationSyncTargetMatchers = computeOrganizationSyncTargetMatchers(options.organizationSyncOptions);
-
-  function removeDevBrowserFromURL(url: URL) {
-    const updatedURL = new URL(url);
-
-    updatedURL.searchParams.delete(constants.QueryParameters.DevBrowser);
-    // Remove legacy dev browser query param key to support local app with v5 using AP with v4
-    updatedURL.searchParams.delete(constants.QueryParameters.LegacyDevBrowser);
-
-    return updatedURL;
-  }
-
-  function buildRedirectToHandshake({ handshakeReason }: { handshakeReason: string }) {
-    const redirectUrl = removeDevBrowserFromURL(authenticateContext.clerkUrl);
-    const frontendApiNoProtocol = authenticateContext.frontendApi.replace(/http(s)?:\/\//, '');
-
-    const url = new URL(`https://${frontendApiNoProtocol}/v1/client/handshake`);
-    url.searchParams.append('redirect_url', redirectUrl?.href || '');
-    url.searchParams.append(
-      constants.QueryParameters.SuffixedCookies,
-      authenticateContext.usesSuffixedCookies().toString(),
-    );
-    url.searchParams.append(constants.QueryParameters.HandshakeReason, handshakeReason);
-
-    if (authenticateContext.instanceType === 'development' && authenticateContext.devBrowserToken) {
-      url.searchParams.append(constants.QueryParameters.DevBrowser, authenticateContext.devBrowserToken);
-    }
-
-    const toActivate = getOrganizationSyncTarget(
-      authenticateContext.clerkUrl,
-      options.organizationSyncOptions,
-      organizationSyncTargetMatchers,
-    );
-    if (toActivate) {
-      const params = getOrganizationSyncQueryParams(toActivate);
-
-      params.forEach((value, key) => {
-        url.searchParams.append(key, value);
-      });
-    }
-
-    return new Headers({ [constants.Headers.Location]: url.href });
-  }
-
-  async function resolveHandshake() {
-    const headers = new Headers({
-      'Access-Control-Allow-Origin': 'null',
-      'Access-Control-Allow-Credentials': 'true',
-    });
-
-    const cookiesToSet: string[] = [];
-
-    if (authenticateContext.handshakeNonce) {
-      // TODO: implement handshake nonce handling, fetch handshake payload with nonce
-    } else {
-      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-      const handshakePayload = await verifyHandshakeToken(authenticateContext.handshakeToken!, authenticateContext);
-      cookiesToSet.push(...handshakePayload.handshake);
-    }
-
-    let sessionToken = '';
-    cookiesToSet.forEach((x: string) => {
-      headers.append('Set-Cookie', x);
-      if (getCookieName(x).startsWith(constants.Cookies.Session)) {
-        sessionToken = getCookieValue(x);
-      }
-    });
-
-    if (authenticateContext.instanceType === 'development') {
-      const newUrl = new URL(authenticateContext.clerkUrl);
-      newUrl.searchParams.delete(constants.QueryParameters.Handshake);
-      newUrl.searchParams.delete(constants.QueryParameters.HandshakeHelp);
-      headers.append(constants.Headers.Location, newUrl.toString());
-      headers.set(constants.Headers.CacheControl, 'no-store');
-    }
-
-    if (sessionToken === '') {
-      return signedOut(authenticateContext, AuthErrorReason.SessionTokenMissing, '', headers);
-    }
-
-    const { data, errors: [error] = [] } = await verifyToken(sessionToken, authenticateContext);
-    if (data) {
-      return signedIn(authenticateContext, data, headers, sessionToken);
-    }
-
-    if (
-      authenticateContext.instanceType === 'development' &&
-      (error?.reason === TokenVerificationErrorReason.TokenExpired ||
-        error?.reason === TokenVerificationErrorReason.TokenNotActiveYet ||
-        error?.reason === TokenVerificationErrorReason.TokenIatInTheFuture)
-    ) {
-      error.tokenCarrier = 'cookie';
-      // This probably means we're dealing with clock skew
-      console.error(
-        `Clerk: Clock skew detected. This usually means that your system clock is inaccurate. Clerk will attempt to account for the clock skew in development.
-
-To resolve this issue, make sure your system's clock is set to the correct time (e.g. turn off and on automatic time synchronization).
-
----
-
-${error.getFullMessage()}`,
-      );
-
-      // Retry with a generous clock skew allowance (1 day)
-      const { data: retryResult, errors: [retryError] = [] } = await verifyToken(sessionToken, {
-        ...authenticateContext,
-        clockSkewInMs: 86_400_000,
-      });
-      if (retryResult) {
-        return signedIn(authenticateContext, retryResult, headers, sessionToken);
-      }
-
-      throw new Error(retryError?.message || 'Clerk: Handshake retry failed.');
-    }
-
-    throw new Error(error?.message || 'Clerk: Handshake failed.');
-  }
+  const handshakeService = new HandshakeService();
 
   async function refreshToken(
     authenticateContext: AuthenticateContext,
@@ -360,21 +224,16 @@ ${error.getFullMessage()}`,
     message: string,
     headers?: Headers,
   ): SignedInState | SignedOutState | HandshakeState {
-    if (isRequestEligibleForHandshake(authenticateContext)) {
-      // Right now the only usage of passing in different headers is for multi-domain sync, which redirects somewhere else.
-      // In the future if we want to decorate the handshake redirect with additional headers per call we need to tweak this logic.
-      const handshakeHeaders = headers ?? buildRedirectToHandshake({ handshakeReason: reason });
+    if (handshakeService.isRequestEligibleForHandshake(authenticateContext)) {
+      const handshakeHeaders =
+        headers ??
+        handshakeService.buildRedirectToHandshake(authenticateContext, organizationSyncTargetMatchers, options, reason);
 
-      // Chrome aggressively caches inactive tabs. If we don't set the header here,
-      // all 307 redirects will be cached and the handshake will end up in an infinite loop.
       if (handshakeHeaders.get(constants.Headers.Location)) {
         handshakeHeaders.set(constants.Headers.CacheControl, 'no-store');
       }
 
-      // Introduce the mechanism to protect for infinite handshake redirect loops
-      // using a cookie and returning true if it's infinite redirect loop or false if we can
-      // proceed with triggering handshake.
-      const isRedirectLoop = setHandshakeInfiniteRedirectionLoopHeaders(handshakeHeaders);
+      const isRedirectLoop = handshakeService.setHandshakeInfiniteRedirectionLoopHeaders(handshakeHeaders);
       if (isRedirectLoop) {
         const msg = `Clerk: Refreshing the session token resulted in an infinite redirect loop. This usually means that your Clerk instance keys do not match - make sure to copy the correct publishable and secret keys from the Clerk dashboard.`;
         console.log(msg);
@@ -465,34 +324,6 @@ ${error.getFullMessage()}`,
     }
   }
 
-  // We want to prevent infinite handshake redirection loops.
-  // We incrementally set a `__clerk_redirection_loop` cookie, and when it loops 3 times, we throw an error.
-  // We also utilize the `referer` header to skip the prefetch requests.
-  function setHandshakeInfiniteRedirectionLoopHeaders(headers: Headers): boolean {
-    if (authenticateContext.handshakeRedirectLoopCounter === 3) {
-      return true;
-    }
-
-    const newCounterValue = authenticateContext.handshakeRedirectLoopCounter + 1;
-    const cookieName = constants.Cookies.RedirectCount;
-    headers.append('Set-Cookie', `${cookieName}=${newCounterValue}; SameSite=Lax; HttpOnly; Max-Age=3`);
-    return false;
-  }
-
-  function handleHandshakeTokenVerificationErrorInDevelopment(error: TokenVerificationError) {
-    // In development, the handshake token is being transferred in the URL as a query parameter, so there is no
-    // possibility of collision with a handshake token of another app running on the same local domain
-    // (etc one app on localhost:3000 and one on localhost:3001).
-    // Therefore, if the handshake token is invalid, it is likely that the user has switched Clerk keys locally.
-    // We make sure to throw a descriptive error message and then stop the handshake flow in every case,
-    // to avoid the possibility of an infinite loop.
-    if (error.reason === TokenVerificationErrorReason.TokenInvalidSignature) {
-      const msg = `Clerk: Handshake token verification failed due to an invalid signature. If you have switched Clerk keys locally, clear your cookies and try again.`;
-      throw new Error(msg);
-    }
-    throw new Error(`Clerk: Handshake token verification failed: ${error.getFullMessage()}.`);
-  }
-
   async function authenticateRequestWithTokenInCookie() {
     const hasActiveClient = authenticateContext.clientUat;
     const hasSessionToken = !!authenticateContext.sessionTokenInCookie;
@@ -503,21 +334,10 @@ ${error.getFullMessage()}`,
      */
     if (authenticateContext.handshakeToken) {
       try {
-        return await resolveHandshake();
+        return await handshakeService.resolveHandshake(authenticateContext);
       } catch (error) {
-        // In production, the handshake token is being transferred as a cookie, so there is a possibility of collision
-        // with a handshake token of another app running on the same etld+1 domain.
-        // For example, if one app is running on sub1.clerk.com and another on sub2.clerk.com, the handshake token
-        // cookie for both apps will be set on etld+1 (clerk.com) so there's a possibility that one app will accidentally
-        // use the handshake token of a different app during the handshake flow.
-        // In this scenario, verification will fail with TokenInvalidSignature. In contrast to the development case,
-        // we need to allow the flow to continue so the app eventually retries another handshake with the correct token.
-        // We need to make sure, however, that we don't allow the flow to continue indefinitely, so we throw an error after X
-        // retries to avoid an infinite loop. An infinite loop can happen if the customer switched Clerk keys for their prod app.
-
-        // Check the handleHandshakeTokenVerificationErrorInDevelopment function for the development case.
         if (error instanceof TokenVerificationError && authenticateContext.instanceType === 'development') {
-          handleHandshakeTokenVerificationErrorInDevelopment(error);
+          handshakeService.handleHandshakeTokenVerificationErrorInDevelopment(error);
         } else {
           console.error('Clerk: unable to resolve handshake:', error);
         }
@@ -817,26 +637,6 @@ export function getOrganizationSyncTarget(
 export type OrganizationSyncTarget =
   | { type: 'personalAccount' }
   | { type: 'organization'; organizationId?: string; organizationSlug?: string };
-
-/**
- * Generates the query parameters to activate an organization or personal account
- * via the FAPI handshake api.
- */
-function getOrganizationSyncQueryParams(toActivate: OrganizationSyncTarget): Map<string, string> {
-  const ret = new Map();
-  if (toActivate.type === 'personalAccount') {
-    ret.set('organization_id', '');
-  }
-  if (toActivate.type === 'organization') {
-    if (toActivate.organizationId) {
-      ret.set('organization_id', toActivate.organizationId);
-    }
-    if (toActivate.organizationSlug) {
-      ret.set('organization_id', toActivate.organizationSlug);
-    }
-  }
-  return ret;
-}
 
 const convertTokenVerificationErrorReasonToAuthErrorReason = ({
   tokenError,
